refactor(bin-designer): extract geometryCore to break facade ↔ sibling cycle
Address PR #1536 review feedback (Greptile P2 + Copilot): every sibling
module (geometryResize/Alignment/Flips) imported runtime helpers from
the geometry.ts facade, while the facade re-exports symbols from those
siblings — a textbook ESM circular dependency that can produce
initialization-order surprises and bundler warnings.

- New `geometryCore.ts` (dependency-free): owns Bounds, rotatePoint,
  getRotatedBounds, computeBounds, getEffectiveBounds, snapToGrid, etc.
- Each sibling now imports only from `./geometryCore`.
- `geometry.ts` shrinks to a pure facade — re-exports from geometryCore
  + geometryResize + geometryAlignment + geometryFlips, no live code.
